Improving Matrix3d.ts: Part 3 (src - line 1700 to 2300)
- improved src code formatting
- improved documentation
- fixed few minor bugs
- moved RotMatrix.test.ts tests to Matrix3d.test.ts to have all Matrix3d
  tests in a single file
