+++ common/changes/@itwin/core-geometry/saeed-torabi-matrix3d-part3_2023-02-06-22-47.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/core-geometry",
+      "comment": "",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/core-geometry"
+}
