+++ core/geometry/src/geometry3d/Point3dVector3d.ts
@@ -942,6 +942,8 @@ export class Vector3d extends XYZ {
    * * It's returning the signed projection magnitude divided by the target magnitude.
    * * To find the projection vector, scale the target vector by the value that this function is returning.
    * * math details can be found at docs/learning/geometry/PointVector.md
+   * * Visualization can be found at https://www.itwinjs.org/sandbox/SaeedTorabi/ProjectVectorOnVector
+   * and https://www.itwinjs.org/sandbox/SaeedTorabi/ProjectVectorOnPlane
    * @param target the target vector
    * @param defaultFraction the returned value in case magnitude square of target vector is very small
    * */
@@ -1350,6 +1352,7 @@ export class Vector3d extends XYZ {
   }
   /**
    * Return the cross product of this vector and vectorB.
+   * * Visualization can be found at https://www.itwinjs.org/sandbox/SaeedTorabi/CrossProduct
    * @param vectorB second vector of cross product
    * @param result optional preallocated result.
    */
